+++ .gitignore
@@ -2,6 +2,7 @@
 .gradle
 /local.properties
 /.idea/caches
+/.idea/deploymentTargetDropDown.xml
 /.idea/libraries
 /.idea/modules.xml
 /.idea/workspace.xml
