+++ app/src/main/java/com/example/tweetssearch/MainApplication.kt
@@ -3,7 +3,7 @@ package com.example.tweetssearch
 import android.app.Application
 import android.content.Context
 import androidx.datastore.preferences.preferencesDataStore
-import com.example.tweetssearch.database.Database
+import com.example.tweetssearch.data.database.Database
 import timber.log.Timber
 
 val Context.dataStore by preferencesDataStore(name = "settings")
