+++ app/src/main/java/com/example/tweetssearch/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.tweetssearch
+package com.example.tweetssearch.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.tweetssearch.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
